cli lib restructuring
